backend/pyproject.toml
[tool.poetry]
name = "flat-digital-backend"
version = "0.1.0"
description = "Flat Digital REST API"
authors = ["Manuel Urgell <manuel.urgell@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
django-cors-headers = "^3.10.1"
djangorestframework = "^3.12.4"
psycopg2 = "^2.9.2"
uWSGI = "^2.0.20"
GitPython = "^3.1.24"

[tool.poetry.dev-dependencies]
django-extensions = "^3.1.5"
flake8 = "^4.0.1"
flake8-import-order = "^0.18.1"
Werkzeug = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
